Convert to float32

+++ benchmark-direct.cpp
@@ -13,7 +13,7 @@
 
 extern "C"
 {
-    extern void dws_conv(double *, double *, double *, double *, int, int, int, int, int, int, int, int, int, int, int, int, double *);
+    extern void dws_conv(float *, float *, float *, float *, int, int, int, int, int, int, int, int, int, int, int, int, float *);
     extern void init_conv(int, int, int, int, int, int, int, int, int, int, int, int, int, int);
 }
 
@@ -21,7 +21,7 @@ extern "C"
 // Helper Functions
 // =================
 // I/O routines
-void save_tensor(std::ofstream &fsave, double *tensor, int size, const char *name)
+void save_tensor(std::ofstream &fsave, float *tensor, int size, const char *name)
 {
     fsave << name << std::endl;
     for (int i = 0; i < size; ++i)
@@ -69,7 +69,7 @@ char *find_string_option(int argc, char **argv, const char *option, char *defaul
 }
 
 // Tensor helper functions
-void fill(double *p, int n, int seed)
+void fill(float *p, int n, int seed)
 {
     static std::random_device rd;
     static std::default_random_engine gen(seed ? seed : rd());
@@ -80,7 +80,7 @@ void fill(double *p, int n, int seed)
     }
 }
 
-void fillDet(double *p, int n)
+void fillDet(float *p, int n)
 {
     for (int i = 0; i < n; ++i)
     {
@@ -88,7 +88,7 @@ void fillDet(double *p, int n)
     }
 }
 
-void fillConst(double *p, int n, double k)
+void fillConst(float *p, int n, float k)
 {
     for (int i = 0; i < n; ++i)
     {
@@ -96,17 +96,17 @@ void fillConst(double *p, int n, double k)
     }
 }
 
-void fillOne(double *p, int n)
+void fillOne(float *p, int n)
 {
     fillConst(p, n, 1.0);
 }
 
-void fillZero(double *p, int n)
+void fillZero(float *p, int n)
 {
     fillConst(p, n, 0.0);
 }
 
-void printTensor(double *p, int B, int W, int H, int C)
+void printTensor(float *p, int B, int W, int H, int C)
 {
     // for (int i = 0; i < n * n; ++i) {
     //     if (i > 0 && i % n == 0) {
@@ -127,7 +127,7 @@ void printTensor(double *p, int B, int W, int H, int C)
             {
                 for (int h = 0; h < H; h += 1)
                 {
-                    double val = *(p + b * img_size + c * mat_size + row_major(h, w, W));
+                    float val = *(p + b * img_size + c * mat_size + row_major(h, w, W));
                     fprintf(stderr, "%f ", val);
                 }
                 fprintf(stderr, "\n");
@@ -228,11 +228,11 @@ void benchmark(bool all_sizes = false)
     int stride_h = 2;
     int stride_w = 2;
 
-    double *input = (double *)calloc(B * C_in * nmax * nmax, sizeof(double));
-    double *F_DW = (double *)calloc(N_dw * C_in * kmax * kmax, sizeof(double));
-    double *F_1D = (double *)calloc(N_1d * C_in * N_dw, sizeof(double));
-    double *output = (double *)calloc(B * C_out * nmax * nmax, sizeof(double));
-    double *depthwise_output = (double *)calloc(B * nmax * nmax * C_in * N_dw, sizeof(double));
+    float *input = (float *)calloc(B * C_in * nmax * nmax, sizeof(float));
+    float *F_DW = (float *)calloc(N_dw * C_in * kmax * kmax, sizeof(float));
+    float *F_1D = (float *)calloc(N_1d * C_in * N_dw, sizeof(float));
+    float *output = (float *)calloc(B * C_out * nmax * nmax, sizeof(float));
+    float *depthwise_output = (float *)calloc(B * nmax * nmax * C_in * N_dw, sizeof(float));
 
     /* For each tensor size */
     int idx = 0;
@@ -316,11 +316,11 @@ void run(int argc, char **argv)
     int stride_h = find_int_arg(argc, argv, "-stride_h", 2);
     int stride_w = find_int_arg(argc, argv, "-stride_w", 2);
 
-    double *input = (double *)calloc(B * C_in * W_in * H_in, sizeof(double));
-    double *F_DW = (double *)calloc(N_dw * C_in * H_f * W_f, sizeof(double));
-    double *F_1D = (double *)calloc(N_1d * C_in * N_dw, sizeof(double));
-    double *output = (double *)calloc(B * C_out * W_out * H_out, sizeof(double));
-    double *depthwise_output = (double *)calloc(B * W_out * H_out * C_in * N_dw, sizeof(double));
+    float *input = (float *)calloc(B * C_in * W_in * H_in, sizeof(float));
+    float *F_DW = (float *)calloc(N_dw * C_in * H_f * W_f, sizeof(float));
+    float *F_1D = (float *)calloc(N_1d * C_in * N_dw, sizeof(float));
+    float *output = (float *)calloc(B * C_out * W_out * H_out, sizeof(float));
+    float *depthwise_output = (float *)calloc(B * W_out * H_out * C_in * N_dw, sizeof(float));
 
     fill(input, B * C_in * W_in * H_in, seed);
     fill(F_DW, N_dw * C_in * H_f * W_f, seed);
