+++ dws-single-blocked.c
@@ -18,7 +18,7 @@ int WIDTH_BLOCK_DW;
 int HEIGHT_FILTER_BLOCK_DW;
 int WIDTH_FILTER_BLOCK_DW;
 
-static void single_op_blocked(double *X, double *F_DW, double *F_1D, double *O, int B, int H_in, int W_in, int C_in, int H_f, int W_f, int N_dw, int H_out, int W_out, int C_out, int stride_h, int stride_w, int b_, int c_, int f_, int h_, int w_)
+static void single_op_blocked(float *X, float *F_DW, float *F_1D, float *O, int B, int H_in, int W_in, int C_in, int H_f, int W_f, int N_dw, int H_out, int W_out, int C_out, int stride_h, int stride_w, int b_, int c_, int f_, int h_, int w_)
 {
     int mat_size = W_in * H_in;
     int f_size = W_f * H_f;
@@ -36,8 +36,8 @@ static void single_op_blocked(double *X, double *F_DW, double *F_1D, double *O, 
     for (int b = 0; b < B_b; b += 1)
     {
         // PTRS TO IMG IN BATCH
-        double *curr_img = X + (b + b_) * img_size;
-        double *curr_out = O + (b + b_) * out_size;
+        float *curr_img = X + (b + b_) * img_size;
+        float *curr_out = O + (b + b_) * out_size;
         for (int c = 0; c < C_b; c += 1)
         {
             for (int f = 0; f < F_b; f += 1)
@@ -47,23 +47,23 @@ static void single_op_blocked(double *X, double *F_DW, double *F_1D, double *O, 
                     for (int w = 0; w < W_b; w += 1)
                     {
                         // Do 2D Convolution channelwise
-                        double *curr_channel = curr_img + mat_size * (c + c_);
+                        float *curr_channel = curr_img + mat_size * (c + c_);
                         // MOST LIKELY SET TO 1.
                         // MICROKERNEL - tile if needed.
                         
-                        double temp = 0.0;
+                        float temp = 0.0;
                         // Depthwise conv
                         for (int w_f = 0; w_f < W_f; w_f += 1)
                         {
                             for (int h_f = 0; h_f < H_f; h_f += 1)
                             {
                                 // PTR TO CURRENT POSITION IN FILTER
-                                double *f_curr = F_DW + f_size * ((c + c_) * N_dw + (f + f_)) + row_major(h_f, w_f, W_f);
+                                float *f_curr = F_DW + f_size * ((c + c_) * N_dw + (f + f_)) + row_major(h_f, w_f, W_f);
 
                                 // PTR TO INPUT POSITION
                                 int h_curr = h_f + stride_h * (h + h_);
                                 int w_curr = w_f + stride_w * (w + w_);
-                                double *curr_inp = curr_channel + row_major(h_curr, w_curr, W_in);
+                                float *curr_inp = curr_channel + row_major(h_curr, w_curr, W_in);
 
                                 // PTR TO INPUT POSITION
                                 temp += *f_curr * *curr_inp;
@@ -73,8 +73,8 @@ static void single_op_blocked(double *X, double *F_DW, double *F_1D, double *O, 
 
                         // LOOP 
                         for (int f_1d = 0; f_1d < C_out; f_1d += 1) {
-                            double *o_curr = curr_out + mat_size_out * f_1d + row_major((h + h_), (w + w_), W_out);
-                            double *curr = F_1D + f_1d * (N_dw * C_in) + N_dw * (c + c_) + (f + f_);
+                            float *o_curr = curr_out + mat_size_out * f_1d + row_major((h + h_), (w + w_), W_out);
+                            float *curr = F_1D + f_1d * (N_dw * C_in) + N_dw * (c + c_) + (f + f_);
                             *o_curr += temp * (*curr);
                         }
                     }
@@ -84,7 +84,7 @@ static void single_op_blocked(double *X, double *F_DW, double *F_1D, double *O, 
     }
 }
 
-static void single_op(double *X, double *F_DW, double *F_1D, double *O, int B, int H_in, int W_in, int C_in, int H_f, int W_f, int N_dw, int H_out, int W_out, int C_out, int stride_h, int stride_w)
+static void single_op(float *X, float *F_DW, float *F_1D, float *O, int B, int H_in, int W_in, int C_in, int H_f, int W_f, int N_dw, int H_out, int W_out, int C_out, int stride_h, int stride_w)
 {
     for (int b = 0; b < B; b += BATCH_BLOCK_DW)
     {
@@ -120,7 +120,7 @@ void init_conv(int bbpw, int fbpw, int wbpw, int hbpw, int cbpw, int bbdw, int c
     WIDTH_FILTER_BLOCK_DW = wfbdw;
 }
 
-void dws_conv(double *X, double *F_DW, double *F_1D, double *O, int B, int H_in, int W_in, int C_in, int H_f, int W_f, int N_dw, int H_out, int W_out, int C_out, int stride_h, int stride_w, double* depthwise_output)
+void dws_conv(float *X, float *F_DW, float *F_1D, float *O, int B, int H_in, int W_in, int C_in, int H_f, int W_f, int N_dw, int H_out, int W_out, int C_out, int stride_h, int stride_w, float* depthwise_output)
 {
     single_op(X, F_DW, F_1D, O, B, H_in, W_in, C_in, H_f, W_f, N_dw, H_out, W_out, C_out, stride_h, stride_w);
 }
